fix: eliminate duplication in Homebrew formulas
Extract shared install/post_install logic into WazuhAgentBundleShared
module. All three formula variants now include the module instead of
duplicating 40 lines each. Reduces duplication from ~87% to 0%.

Also fix test blocks to check actual installed artifacts instead of
a non-existent binary.
